Final changes
This is not completely 100% correct, this was a challenge for me using
Github for the first time and realizing my previous UI work has not
been as intense and really pushed my skills to grow.

Added overlay for help link to display the help links from the
help.json file.