fixing duplicate json object name for item 3 in created json object
im not seeing a way to validate json via qunit. this is easily checked by any json validator though. just have to remove the 'var smithFamily = ' substring at the front. I left this on because I was getting a script error failure when just placing a strict json object in the file with no identification