cleaning up logic placement, logging, and unnecessary text
i was hitting an issue where since the jsanddom.js file is included in both the JsAndDom.html file and the test-runner html file, I couldn't explicitly call each of the DOM manipulation funcs inside the controller file. I had to do inline-calling of the js funcs inside of the JsAndDom.html file so that the unit tests wouldn't break.