final touches, testing
had some final styles to apply, some icons to add, had to fiddle with the js a bit. I did some testing in latest versions of Firefox and Chrome, and fixed a bug I saw in the js for Firefox. Getting ready to set up bootcamp on this machine so I can test the code in IE.